+++ lib/default.nix
@@ -1,4 +1,4 @@
-{ nixpkgs-lib }:
+{ lib }:
 rec {
   hypervisors = [
     "qemu"
@@ -7,6 +7,7 @@ rec {
     "crosvm"
     "kvmtool"
     "stratovirt"
+    "alioth"
   ];
 
   hypervisorsWithNetwork = hypervisors;
@@ -24,14 +25,15 @@ rec {
       fst // {
         letter = snd;
       }
-    ) (nixpkgs-lib.zipLists volumes (
-      nixpkgs-lib.drop offset nixpkgs-lib.strings.lowerChars
+    ) (lib.zipLists volumes (
+      lib.drop offset lib.strings.lowerChars
     ));
 
   createVolumesScript = pkgs: pkgs.lib.concatMapStringsSep "\n" (
     { image
     , label
     , size ? throw "Specify a size for volume ${image} or use autoCreate = false"
+    , mkfsExtraArgs
     , fsType ? defaultFsType
     , autoCreate ? true
     , ...
@@ -39,7 +41,7 @@ rec {
       (label != null && !autoCreate) "Volume is not automatically labeled unless autoCreate is true. Volume has to be labeled manually, otherwise it will not be identified"
       (let labelOption =
              if autoCreate then
-               (if builtins.elem fsType ["ext2" "ext3" "ext4" "xfs"] then "-L"
+               (if builtins.elem fsType ["ext2" "ext3" "ext4" "xfs" "btrfs"] then "-L"
                 else if fsType == "vfat" then "-n"
                 else (pkgs.lib.warnIf (label != null)
                   "Will not label volume ${label} with filesystem type ${fsType}. Open an issue on the microvm.nix project to request a fix."
@@ -48,15 +50,19 @@ rec {
            labelArgument =
              if (labelOption != null && label != null) then "${labelOption} '${label}'"
              else "";
-      in (nixpkgs-lib.optionalString autoCreate ''
-      PATH=$PATH:${with pkgs.buildPackages; lib.makeBinPath [ coreutils util-linux e2fsprogs ]}
+             mkfsExtraArgsString =
+              if mkfsExtraArgs != null
+              then lib.escapeShellArgs mkfsExtraArgs
+              else " ";
+      in (lib.optionalString autoCreate ''
+      PATH=$PATH:${with pkgs.buildPackages; lib.makeBinPath [ coreutils util-linux e2fsprogs xfsprogs dosfstools btrfs-progs ]}
 
       if [ ! -e '${image}' ]; then
         touch '${image}'
         # Mark NOCOW
         chattr +C '${image}' || true
-        fallocate -l${toString size}MiB '${image}'
-        mkfs.${fsType} ${labelArgument} '${image}'
+        truncate -s ${toString size}M '${image}'
+        mkfs.${fsType} ${labelArgument} ${mkfsExtraArgsString} '${image}'
       fi
     '')));
 
@@ -64,7 +70,6 @@ rec {
 
   makeMacvtap = { microvmConfig, hypervisorConfig }:
     import ./macvtap.nix {
-      inherit microvmConfig hypervisorConfig;
-      lib = nixpkgs-lib;
+      inherit microvmConfig hypervisorConfig lib;
     };
 }